+++ tests/features/union/union.c
@@ -95,7 +95,7 @@ next_2:;
     match_ = 0;
     goto end_match_1;
 next_3:;
-    exit(1);
+    milone_exhaust_error("union/union.milone", 27, 4);
 end_match_1:;
     variant_2 = (struct Limit_){.discriminant = Limit_LimitVal, .Limit_LimitVal = 1};
     if ((variant_2.discriminant != Limit_LimitVal)) goto next_5;
@@ -107,7 +107,7 @@ next_5:;
     if ((variant_2.discriminant != Limit_LimitDiv)) goto next_6;
     exit(2);
 next_6:;
-    exit(1);
+    milone_exhaust_error("union/union.milone", 32, 4);
 end_match_4:;
     switch ((struct ApiResponse_){.discriminant = ApiResponse_ARCancel}.discriminant) {
         case ApiResponse_ARCancel:
@@ -161,7 +161,7 @@ end_match_10:;
             goto clause_18;
 
         default:
-            exit(1);
+            milone_exhaust_error("union/union.milone", 53, 2);
     }
 clause_17:;
     switch_1 = 0;